fix(xurl): resolve recall_cmd module path for test compilation (#1623)
cli_xurl.rs is `#[path]`-included from cli.rs, which is in turn
`#[path]`-included from integration tests (tests/e2e.rs,
tests/review_mode_compat.rs). Those test binaries do not declare
`mod recall_cmd;` in their crate root, so any `crate::recall_cmd::*`
reference from cli_xurl.rs fails to resolve.

Move the recall dispatch into a new top-level module `xurl_cmd.rs`
declared only from main.rs (following the existing `<name>_cmd.rs`
pattern alongside recall_cmd.rs). cli_xurl.rs now contains only the
clap argument definitions and the self-contained `handle_threads`
helper; the binary dispatches `Commands::Xurl` via `xurl_cmd::handle_xurl`.

Also drop a redundant WHAT-only comment from main.rs to keep the
module under the 8000-token monolith threshold after the new
`mod xurl_cmd;` declaration.
